+++ test/test_taskcraft.py
@@ -37,6 +37,7 @@ class TaskCraftPipeline():
         
 
         ### Either use DepthQAGenerator or WidthQAGenerator
+        
         # self.task_step3 = WidthQAGenerator(
         #     llm_serving = llm_serving
         # )
